requirements.txt
asgiref==3.8.1
cachetools==6.2.1
certifi==2025.4.26
charset-normalizer==3.4.2
django==5.2.4
google-api-core==2.26.0
google-api-python-client==2.184.0
google-auth==2.41.1
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2
googleapis-common-protos==1.70.0
gunicorn==23.0.0
httplib2==0.31.0
idna==3.10
oauthlib==3.3.1
packaging==25.0
proto-plus==1.26.1
protobuf==6.33.0
pyasn1==0.6.1
pyasn1-modules==0.4.2
pyparsing==3.2.5
pytils==0.4.3
requests==2.32.4
requests-oauthlib==2.0.0
rsa==4.9.1
six==1.17.0
sqlparse==0.5.3
typing-extensions==4.13.2
uritemplate==4.2.0
urllib3==2.4.0
